wav: relax Decode input to support io.Reader without Close method
Calling Close on the returned beep.ReadSeekCloser is a nop on the
underlying io.Reader the supplied reader to Decode is lacking a
Close method.

The motivating reason to relax the input parameter to Decode is
to allow for bytes.Buffer and bufio.Reader to be used, both of
which are lacking a Close method.